create a fileinputfield with a handler to be called
